refactor(handlers): collapse the duplicated OpenFGA plumbing and scope-handler skeleton
Closes #879.

Seven classes carried a `getFgaClient()` in two variants. Six were a
byte-identical private pair (`isFgaClientAvailable()` + a memoizing
`getFgaClient()`); the seventh, in ResolvesOutboxTooling, rebuilt the client —
and therefore its Guzzle connection pool — on every call and offered no
injection seam. All of them now compose Concerns\ResolvesFgaClient, so the
outbox path gains memoization and there is one place left to change how the
client is built.

On top of that, the three scope endpoints (/auth/admin-scopes, /auth/test-scopes,
/auth/dashboard-scopes) shared ~60 lines of triplicated plumbing: method and
Accept validation, CORS, the no-store directive, extracting `sub` from the OIDC
token with both UnauthorizedException branches, reading the global-admin role,
and gating on whether OpenFGA is reachable. A correction to the fail-closed
contract in one of them was a correction in one of them. AbstractScopesHandler
now owns that sequence in a `final handle()` and delegates to a single hook:

    buildScopePayload(string $sub, bool $isGlobalAdmin, ?ResourceAdminService $service): array

A null $service is precisely the "OpenFGA unavailable" case, which is where each
subclass already had its fail-closed literal. The three handlers keep their
docblocks and shrink to that one method.

No behavior change: response shapes, key order and the per-type fail-closed
semantics of #793 are untouched, and the existing handler tests pass unmodified
(227 tests in Handlers/Auth + Handlers/Admin + ResourceAdminServiceTest, plus 83
covering the two ResolvesOutboxTooling consumers). PHPStan level 10 and phpcs
clean.
